Krisz code

+++ Project/e_Sport_Projekt/backend/controllers/gameController.js
@@ -15,6 +15,10 @@ const gameList = async (req, res) => {
 const gameUpdate = async (req, res) => {
     const { id, name_get, name_set } = req.body;
 
+    if(!id || !name_get || !name_set){
+        return res.status(400).json({message: "Hiányos adatok!"});
+    }
+
     try {
 
         const existingGame = await prisma.games.findFirst({
@@ -23,14 +27,19 @@ const gameUpdate = async (req, res) => {
             }
         });
 
-        if (existingGame || existingGame.name == req.body.name_set) {
+        if(!existingGame){
+            return res.status(400).json({message: "Nincs ilyen játék!"});
+        }
+
+        if (existingGame.name == name_set) {
             return res.status(400).json({ message: "Ilyen nevű játék már létezik!" })
         };
 
         const game = await prisma.games.update({
             where: {
-                id: id,
-                name: name_get
+                id: existingGame.id,
+                name: existingGame.name
+                
             },
             data: {
                 name: name_set
@@ -47,6 +56,10 @@ const gameUpdate = async (req, res) => {
 const gameInsert = async (req, res) => {
     const { name } = req.body;
 
+    if(!name || name==""){
+        return res.status(400).json({message: "Hiányos adat!"})
+    }
+
     try {
         const existingGame = await prisma.games.findFirst({
             where: {
@@ -54,12 +67,13 @@ const gameInsert = async (req, res) => {
             }
         });
 
+        
         if (existingGame) {
             return res.status(400).json({ message: "Hiba! Ilyen nevű játék már létezik!" })
         }
 
         const game = await prisma.games.create({ data: { name: name } })
-        return res.status(200).json({ message: "Sikeres adatfrissítés!" })
+        return res.status(200).json({ message: "Játék felvéve!" })
 
     }
     catch (err) {
@@ -73,11 +87,24 @@ const gameDelete = async (req, res) => {
     const { id } = req.body;
 
     try {
+
+        const tournaments = await prisma.tournaments.findFirst({
+            where:{
+                gae_id: id
+            }
+        })
+
+        if(tournaments){
+            return res.status(400).json({message: "Nem törölhetsz olyan játékot, amihez van verseny hozzárendelve!"});
+        }
+
+
         const game = await prisma.games.delete({
             where: {
                 id: id
             }
         })
+
         res.status(200).json({ message: "Sikeres törlés!" })
 
     } catch (err) {
